// Develop/src/interfaces/Candidate.interface.tsx
// TODO: Create an interface for the Candidate objects returned by the API
export interface Candidate {
    name: string;          
    username: string;
    location: string;     
    email: string;        
    avatar: string            
    avatar_url: string;   
    html_url: string;              
    company: string; 
}